refactor(settings): move bash-approval migration out of src/shared
migrateLegacyGlobalBashApprovalOverride() was a stateful, host-triggered
one-shot migration workflow (config reads/writes, retry-on-failure,
logging) living in src/shared/settingsView/handlers/approvalHandlers.ts,
a module documented and tested (SharedSettingsViewBoundary.vitest.ts) as
pure message builders shared between hosts. CLAUDE.md is explicit that
host-neutral orchestration belongs in src/controllers/, not src/shared/
— the boundary test only greps import specifiers, so this embedded-logic
violation slipped through undetected.

Move the migration (and its BASH_APPROVAL_CONFIG_TARGET constant) into
a new src/controllers/settingsView/BashApprovalGlobalMigration.ts,
following the existing function-style controller convention (e.g.
HistoryActionOutcomes.ts, ProfileMessageBuilder.ts). approvalHandlers.ts
now only contains buildApprovalSettingsMessage, its actual shared
concern. Update the extension.ts call site and relocate the two test
files to test-kernel/controllers to match.

No behavior change.
